apibara_backend/pyproject.toml
[tool.poetry]
name = "isaac_api"
version = "0.1.0"
description = ""
authors = ["Francesco Ceccon <francesco@ceccon.me>"]

[tool.poetry.scripts]
isaac-api = "isaac_api.cli:cli"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
grpcio = "^1.46.3"
aiochannel = "^1.1.1"

[tool.poetry.dev-dependencies]
grpcio-tools = "^1.46.3"
pytest = "^5.2"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
